requirements.txt
Flask
folium
pyyaml
